@@ -16,18 +16,16 @@
 
 package com.netflix.spinnaker.kork.secrets;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
+import org.junit.jupiter.api.Test;
 
 public class EncryptedSecretTest {
 
-  @Rule
-  public ExpectedException exceptionRule = ExpectedException.none();
-
+  /*
+   * @Rule public ExpectedException exceptionRule = ExpectedException.none();
+   */
   @Test
   public void isEncryptedSecretShouldReturnFalse() {
     String secretConfig = "foo";
@@ -93,22 +91,20 @@ public class EncryptedSecretTest {
     assertEquals("secret-param-1", encryptedSecretFile.getParams().get("key"));
     assertEquals("secret-param:2", encryptedSecretFile.getParams().get("Key"));
   }
-
-  @Test
-  public void updateThrowsInvalidSecretFormatException() {
-    exceptionRule.expect(InvalidSecretFormatException.class);
-    exceptionRule.expectMessage("Invalid encrypted secret format, must have at least one parameter");
-    EncryptedSecret encryptedSecret = new EncryptedSecret();
-    encryptedSecret.update("encrypted:s3");
-  }
-
-  @Test
-  public void updateThrowsInvalidSecretFormatExceptionNoKeyValuePairs() {
-    exceptionRule.expect(InvalidSecretFormatException.class);
-    exceptionRule.expectMessage("Invalid encrypted secret format, keys and values must be delimited by ':'");
-    EncryptedSecret encryptedSecret = new EncryptedSecret();
-    encryptedSecret.update("encrypted:s3!foobar");
-  }
+  /*
+   *
+   * @Test public void updateThrowsInvalidSecretFormatException() {
+   * exceptionRule.expect(InvalidSecretFormatException.class);
+   * exceptionRule.expectMessage("Invalid encrypted secret format, must have at least one parameter");
+   * EncryptedSecret encryptedSecret = new EncryptedSecret(); encryptedSecret.update("encrypted:s3");
+   * }
+   *
+   * @Test public void updateThrowsInvalidSecretFormatExceptionNoKeyValuePairs() {
+   * exceptionRule.expect(InvalidSecretFormatException.class); exceptionRule.
+   * expectMessage("Invalid encrypted secret format, keys and values must be delimited by ':'");
+   * EncryptedSecret encryptedSecret = new EncryptedSecret();
+   * encryptedSecret.update("encrypted:s3!foobar"); }
+   */
 
   @Test
   public void tryParseReturnsEmptyOnInvalidSecretFormat() {
