@@ -1,13 +1,13 @@
 package com.netflix.spinnaker.kork.secrets;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -46,7 +46,7 @@ public class SecretBeanPostProcessorTest {
     }
   };
 
-  @Before
+  @BeforeAll
   public void setup() {
     mutablePropertySources.addLast(propertySource);
     mutablePropertySources.addLast(enumerablePropertySource);
