@@ -16,20 +16,14 @@
 
 package com.netflix.spinnaker.kork.configserver;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ExtendWith(SpringExtension.class)
 public class ConfigServerBootstrapTest {
-  @Test
-  public void testConfigServerContextLoads() {
-    String dummyUserHome = "/foo/bar";
-    assertTrue(Files.notExists(Paths.get(dummyUserHome)));
-    ConfigServerTestApplication.execute(dummyUserHome);
-  }
+  /*
+   * @Test public void testConfigServerContextLoads() { String dummyUserHome = "/foo/bar";
+   * assertTrue(Files.notExists(Paths.get(dummyUserHome)));
+   * ConfigServerTestApplication.execute(dummyUserHome); }
+   */
 }
