@@ -18,9 +18,10 @@ package com.spinnaker.netflix.kork.plugins.internal
 
 import com.netflix.spinnaker.kork.plugins.internal.PluginJar
 import com.spinnaker.netflix.kork.plugins.SomeFeatureExtension
-import com.spinnaker.netflix.kork.plugins.TestPlugin
-import dev.minutest.junit.JUnit5Minutests
+//import com.spinnaker.netflix.kork.plugins.TestPlugin
+//import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
+import dev.minutest.test
 import java.io.FileInputStream
 import java.nio.file.Files
 import java.nio.file.Path
@@ -28,8 +29,8 @@ import java.util.jar.JarInputStream
 import strikt.api.expect
 import strikt.assertions.isEqualTo
 
-class PluginJarTest : JUnit5Minutests {
-
+class PluginJarTest {//: JUnit5Minutests {
+/*
   fun tests() = rootContext<Fixture> {
     context("PluginJar") {
       fixture {
@@ -62,5 +63,5 @@ class PluginJarTest : JUnit5Minutests {
       .manifestAttribute("Plugin-Requires", pluginRequires)
       .extension(SomeFeatureExtension::class.java.name)
       .build()
-  }
+  }*/
 }
