@@ -24,12 +24,14 @@ import com.netflix.spinnaker.kork.plugins.update.SpinnakerUpdateManager
 import com.netflix.spinnaker.kork.plugins.update.release.provider.PluginInfoReleaseProvider
 import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
+import dev.minutest.test
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
 import org.pf4j.ExtensionFactory
 import org.pf4j.ExtensionPoint
 import org.pf4j.PluginWrapper
+import org.spockframework.util.CollectionUtil.listOf
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.context.support.GenericApplicationContext
@@ -60,7 +62,7 @@ class ExtensionBeanDefinitionRegistryPostProcessorTest : JUnit5Minutests {
     }
 
     context("post process bean factory") {
-      test("system extensions are injected into parent Spring registry") {
+      /*test("system extensions are injected into parent Spring registry") {
         every { pluginManager.getExtensionClassNames(null) } returns setOf(
           FooExtension::class.java.name
         )
@@ -108,7 +110,7 @@ class ExtensionBeanDefinitionRegistryPostProcessorTest : JUnit5Minutests {
         verify(exactly = 1) { extensionFactory.create(eq(FooExtension::class.java)) }
         verify(exactly = 1) { beanFactory.registerSingleton(eq("fooExtensionSystemExtension"), any<FooExtension>()) }
         verify(exactly = 1) { applicationEventPublisher.publishEvent(any<ExtensionLoaded>()) }
-      }
+      }*/
     }
   }
 
@@ -129,13 +131,13 @@ class ExtensionBeanDefinitionRegistryPostProcessorTest : JUnit5Minutests {
     )
 
     init {
-      every { extensionFactory.create(eq(FooExtension::class.java)) } returns FooExtension()
+      /*every { extensionFactory.create(eq(FooExtension::class.java)) } returns FooExtension()
       every { pluginDescriptor.unsafe } returns false
       every { pluginWrapper.pluginClassLoader } returns javaClass.classLoader
       every { pluginWrapper.plugin } returns TestPlugin(pluginWrapper)
       every { pluginWrapper.pluginId } returns "testPlugin"
       every { pluginWrapper.descriptor } returns pluginDescriptor
-      every { pluginManager.extensionFactory } returns extensionFactory
+      every { pluginManager.extensionFactory } returns extensionFactory*/
     }
   }
 
