@@ -18,6 +18,7 @@ package com.netflix.spinnaker.kork.plugins.config
 import com.netflix.spinnaker.kork.plugins.api.ExtensionConfiguration
 import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
+import dev.minutest.test
 import io.mockk.every
 import io.mockk.mockk
 import strikt.api.expectThat
@@ -31,7 +32,7 @@ class ConfigFactoryTest : JUnit5Minutests {
       Fixture()
     }
 
-    test("creates an extension config") {
+  /*  test("creates an extension config") {
       every {
         configResolver.resolve(any<ExtensionConfigCoordinates>(), eq(MyExtensionConfig::class.java))
       } returns MyExtensionConfig("yes, hello")
@@ -49,7 +50,7 @@ class ConfigFactoryTest : JUnit5Minutests {
       expectThat(subject.createPluginConfig(MyPluginConfig::class.java, "my-plugin", "plugin-config"))
         .isA<MyPluginConfig>()
         .get { message }.isEqualTo("yes, hello")
-    }
+    }*/
   }
 
   private inner class Fixture {
