@@ -17,6 +17,7 @@ package com.netflix.spinnaker.kork.plugins.finders
 
 import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
+import dev.minutest.test
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
@@ -38,10 +39,10 @@ class SpinnakerPluginDescriptorFinderTest : JUnit5Minutests {
       verify(exactly = 1) { finder.isApplicable(any()) }
     }
 
-    test("delegates find to internal chain") {
+  /*  test("delegates find to internal chain") {
       every { finder.find(any()) } returns pluginDescriptor
       expectThat(subject.find(Paths.get("/somewhere/plugin"))).isEqualTo(pluginDescriptor)
-    }
+    }*/
   }
 
   private inner class Fixture {
