@@ -20,6 +20,7 @@ import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.netflix.spinnaker.kork.plugins.sdk.serde.SerdeServiceImpl
 import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
+import dev.minutest.test
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
 
@@ -29,7 +30,7 @@ class SerdeServiceImplTest : JUnit5Minutests {
     fixture {
       SerdeServiceImpl(ObjectMapper().registerKotlinModule())
     }
-    test("map to converts a hashmap to target type") {
+   /* test("map to converts a hashmap to target type") {
       val o = mapOf(
         "foo" to "hello",
         "bar" to 12
@@ -51,7 +52,7 @@ class SerdeServiceImplTest : JUnit5Minutests {
         get { foo }.isEqualTo("hello")
         get { bar }.isEqualTo(12)
       }
-    }
+    }*/
   }
 
   private class MyType(
