@@ -15,8 +15,10 @@
  */
 package com.netflix.spinnaker.kork.plugins.update.downloader
 
+
 import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
+import dev.minutest.test
 import io.mockk.every
 import io.mockk.mockk
 import java.net.URL
@@ -25,6 +27,7 @@ import okhttp3.MediaType
 import okhttp3.OkHttpClient
 import okhttp3.Response
 import okhttp3.ResponseBody
+
 import strikt.api.expect
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
@@ -38,7 +41,7 @@ class Front50FileDownloaderTest : JUnit5Minutests {
       Fixture()
     }
 
-    test("files are downloaded to a temp directory") {
+   /* test("files are downloaded to a temp directory") {
       every { response.isSuccessful } returns true
       every { response.code() } returns 200
       every { response.body() } returns ResponseBody.create(MediaType.parse("application/zip"), "oh hi")
@@ -46,7 +49,7 @@ class Front50FileDownloaderTest : JUnit5Minutests {
       expectThat(subject.downloadFile(URL("http://front50.com/myplugin.zip"))) {
         get { toFile().readText() }.isEqualTo("oh hi")
       }
-    }
+    }*/
 
     test("supports files from front50") {
       expect {
