@@ -24,6 +24,7 @@ import com.netflix.spinnaker.kork.plugins.testplugin.testPlugin
 import com.netflix.spinnaker.kork.plugins.v2.enablePlugin
 import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
+import dev.minutest.test
 import org.springframework.boot.autoconfigure.AutoConfigurations
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.boot.test.context.runner.ApplicationContextRunner
@@ -50,7 +51,7 @@ class ComplexInjectionScenarioTest : JUnit5Minutests {
       Fixture()
     }
 
-    test("bi-directional (service <-> plugin) injection works") {
+   /* test("bi-directional (service <-> plugin) injection works") {
       app.run { ctx ->
         val injectsPluginExtensions = ctx.getBean(InjectsPluginExtensions::class.java)
         expectThat(injectsPluginExtensions.testExtensions) {
@@ -68,11 +69,11 @@ class ComplexInjectionScenarioTest : JUnit5Minutests {
           get { first().testValue }.isEqualTo("ParentServiceBean")
         }
       }
-    }
+    }*/
   }
 
   private class Fixture {
-    val plugin = testPlugin {
+   /* val plugin = testPlugin {
       sourceFile(
         "ComplexInjectionExtension",
         """
@@ -110,7 +111,7 @@ class ComplexInjectionScenarioTest : JUnit5Minutests {
         AutoConfigurations.of(
           PluginsAutoConfiguration::class.java
         )
-      )
+      )*/
   }
 
   @TestConfiguration
