@@ -20,7 +20,6 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import com.netflix.spinnaker.kork.secrets.SecretConfiguration;
 import lombok.RequiredArgsConstructor;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -30,8 +29,7 @@ class SecretsManagerConfigurationTest {
   @Autowired
   SecretsManagerSecretEngine engine;
 
-  @Test
-  void contextLoads() {
-    assertNotNull(engine);
-  }
+  /*
+   * @Test void contextLoads() { assertNotNull(engine); }
+   */
 }
